Add new template, css style, controller...

+++ crm/src/main/java/com/bm/rra/crm/repository/SalesManagerRepository.java
@@ -6,6 +6,6 @@ import com.bm.rra.crm.repository.base.BaseRepository;
 
 public interface SalesManagerRepository extends BaseRepository<SalesManager, Long> {
 
-    SalesManager getById(Integer id);
+    SalesManager getById(Long id);
 
 }
